% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brinley.R
\name{brinley}
\alias{brinley}
\title{Modified Brinley's plots}
\usage{
brinley(data, formula, facet = FALSE, colour = NULL, background = TRUE,
  CI = TRUE)
}
\arguments{
\item{data}{A dataframe containing all relevant variables}

\item{formula}{A simple formula of the form y ~ x (nb: with only one predictor)}

\item{facet}{A facet... (should be set to NULL otherwise)}

\item{colour}{A colour... (should be set to NULL otherwise)}

\item{background}{Should the whole dataset be plotted in background ?}

\item{CI}{Should it plot the confidence intervals ?}
}
\description{
Plot...
}
\examples{
\dontrun{
data(brinley_data)

brinley(
data, formula = outcome ~ session, facet = condition, colour = pain,
background = TRUE, CI = TRUE)
}

}
\references{
Blampied, N. M. (2017). Analyzing Therapeutic Change Using
Modified Brinley Plots: History, Construction, and Interpretation. Behavior
Therapy, 48, 115-117.
}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
