% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brinley_data.R
\docType{data}
\name{brinley_data}
\alias{brinley_data}
\title{Data for brinley...}
\format{A data frame with ... rows and ... variables:
\describe{
  \item{participant}{...}
  \item{session}{...}
  \item{condition}{...}
  \item{outcome}{...}
  \item{pain}{...}
}}
\usage{
brinley_data
}
\description{
A dataset containing ...
}
\keyword{datasets}
