% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\docType{data}
\name{dominance}
\alias{dominance}
\title{Data for ...}
\format{A data frame with ... rows and ... variables:
\describe{
  \item{exp}{...}
  \item{group}{...}
  \item{value}{...}
  \item{ppt}{...}
}}
\usage{
dominance
}
\description{
A dataset containing ...
}
\keyword{datasets}
