% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howell.R
\docType{data}
\name{howell}
\alias{howell}
\title{Data for ...}
\format{A data frame with ... rows and ... variables:
\describe{
  \item{height}{...}
  \item{weight}{...}
  \item{age}{...}
  \item{male}{...}
}}
\usage{
howell
}
\description{
A dataset containing ...
}
\keyword{datasets}
