% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_to_gamma.R
\name{normal_to_gamma}
\alias{normal_to_gamma}
\title{Converting a Gaussian distribution to a Gamma distribution}
\usage{
normal_to_gamma(mu, sigma)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard deviation}
}
\description{
Plotting two distributions. A Normal distribution and a Gamma distribution
with same mean and standard deviation.
}
\examples{
normal_to_gamma(mu = 50, sigma = 10)

}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
