% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_priors.R
\name{prior_scales}
\alias{prior_scales}
\title{Converting prior from logit scale to normal scale}
\usage{
prior_scales(prior, ...)
}
\arguments{
\item{prior}{A prior in character strings}

\item{...}{Additionnal parameters to be passed to plot}
}
\description{
Plot...
}
\examples{
prior_scales(prior = "exp(2)")
prior_scales(prior = "norm(0,5)")
prior_scales(prior = "halfcauchy(5)")

}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
