% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Plotting p-value distribution}
\usage{
pdist(cohensd = 0, n = 50, nsims = 10000, alpha = 0.05)
}
\arguments{
\item{cohensd}{Standardised effect size}

\item{n}{Sample size (per group)}

\item{nsims}{Numer of simulations}

\item{alpha}{Alpha level}
}
\description{
Plot...
}
\examples{
pdist(cohensd = 0.5, n = 50, nsims = 1e3, alpha = .05)

}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
